+++ pkg/controller/utils/clusterinfo.go
@@ -3,18 +3,16 @@ package utils
 import (
 	"context"
 	"fmt"
-	"net/url"
-	"strings"
 
-	configv1 "github.com/openshift/api/config/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	machineapi "github.com/openshift/machine-api-operator/pkg/apis/machine/v1beta1"
 	awsproviderapi "sigs.k8s.io/cluster-api-provider-aws/pkg/apis/awsproviderconfig/v1beta1"
 
 	"sigs.k8s.io/yaml"
+
+	baseutils "github.com/openshift/cloud-ingress-operator/pkg/utils"
 )
 
 // installConfig represents the bare minimum requirement to get the AWS cluster region from the install-config
@@ -27,40 +25,6 @@ type installConfig struct {
 	} `json:"platform"`
 }
 
-const masterMachineLabel string = "machine.openshift.io/cluster-api-machine-role"
-
-// GetClusterBaseDomain returns the installed cluster's base domain name
-func GetClusterBaseDomain(kclient client.Client) (string, error) {
-	infra, err := getInfrastructureObject(kclient)
-	if err != nil {
-		return "", err
-	}
-	// This starts with "api." that needs to be removed.
-	u, err := url.Parse(infra.Status.APIServerURL)
-	if err != nil {
-		return "", fmt.Errorf("Couldn't parse the cluster's URI from %s: %s", infra.Status.APIServerURL, err)
-	}
-	return u.Hostname()[4:], nil
-}
-
-// GetClusterPlatform will return the installed cluster's platform type
-func GetClusterPlatform(kclient client.Client) (string, error) {
-	infra, err := getInfrastructureObject(kclient)
-	if err != nil {
-		return "", err
-	}
-	return string(infra.Status.Platform), nil
-}
-
-// GetClusterName returns the installed cluster's name (max 27 characters)
-func GetClusterName(kclient client.Client) (string, error) {
-	infra, err := getInfrastructureObject(kclient)
-	if err != nil {
-		return "", err
-	}
-	return infra.Status.InfrastructureName, nil
-}
-
 // GetMasterNodeSubnets returns all the subnets for Machines with 'master' label.
 // return structure:
 // {
@@ -70,7 +34,7 @@ func GetClusterName(kclient client.Client) (string, error) {
 //
 func GetMasterNodeSubnets(kclient client.Client) (map[string]string, error) {
 	subnets := make(map[string]string)
-	machineList, err := GetMasterMachines(kclient)
+	machineList, err := baseutils.GetMasterMachines(kclient)
 	if err != nil {
 		return subnets, err
 	}
@@ -94,7 +58,7 @@ func GetMasterNodeSubnets(kclient client.Client) (map[string]string, error) {
 
 	// Infra object gives us the Infrastructure name, which is the combination of
 	// cluster name and identifier.
-	infra, err := getInfrastructureObject(kclient)
+	infra, err := baseutils.GetInfrastructureObject(kclient)
 	if err != nil {
 		return subnets, err
 	}
@@ -106,7 +70,7 @@ func GetMasterNodeSubnets(kclient client.Client) (map[string]string, error) {
 
 // GetClusterRegion returns the installed cluster's AWS region
 func GetClusterRegion(kclient client.Client) (string, error) {
-	infra, err := getInfrastructureObject(kclient)
+	infra, err := baseutils.GetInfrastructureObject(kclient)
 	if err != nil {
 		return "", err
 	} else if infra.Status.PlatformStatus == nil {
@@ -114,68 +78,6 @@ func GetClusterRegion(kclient client.Client) (string, error) {
 		return readClusterRegionFromConfigMap(kclient)
 	}
 	return infra.Status.PlatformStatus.AWS.Region, nil
-}
-
-// GetMasterNodes returns a machineList object whose .Items can be iterated
-// over to perform actions on/with information from each master machine object
-func GetMasterMachines(kclient client.Client) (*machineapi.MachineList, error) {
-	machineList := &machineapi.MachineList{}
-	listOptions := []client.ListOption{
-		client.InNamespace("openshift-machine-api"),
-		client.MatchingLabels{masterMachineLabel: "master"},
-	}
-	err := kclient.List(context.TODO(), machineList, listOptions...)
-	if err != nil {
-		return nil, err
-	}
-	return machineList, nil
-}
-
-// GetClusterMasterInstancesIDs gets all the instance IDs for Master nodes
-// For AWS the form is aws:///<availability zone>/<instance ID>
-// This could come from parsing the arbitrarily formatted .Status.ProviderStatus
-// but .Spec.ProviderID is standard
-func GetClusterMasterInstancesIDs(kclient client.Client) ([]string, error) {
-	machineList, err := GetMasterMachines(kclient)
-	if err != nil {
-		return []string{}, err
-	}
-
-	ids := make([]string, 0)
-
-	for _, machineObj := range machineList.Items {
-		r := strings.LastIndex(*machineObj.Spec.ProviderID, "/")
-		if r != -1 {
-			n := *machineObj.Spec.ProviderID
-			ids = append(ids, n[r+1:])
-		}
-	}
-	return ids, nil
-}
-
-func getInfrastructureObject(kclient client.Client) (*configv1.Infrastructure, error) {
-	infra := &configv1.Infrastructure{}
-	ns := types.NamespacedName{
-		Namespace: "",
-		Name:      "cluster",
-	}
-	err := kclient.Get(context.TODO(), ns, infra)
-	if err != nil {
-		return nil, err
-	}
-	return infra, nil
-}
-
-// AWSOwnerTag returns owner taglist for the cluster
-func AWSOwnerTag(kclient client.Client) (map[string]string, error) {
-	m := make(map[string]string)
-	name, err := GetClusterName(kclient)
-	if err != nil {
-		return m, err
-	}
-
-	m[fmt.Sprintf("kubernetes.io/cluster/%s", name)] = "owned"
-	return m, nil
 }
 
 func readClusterRegionFromConfigMap(kclient client.Client) (string, error) {
